Refactor and fix array and span libfuncs (#461)
* Rewrite some array libfuncs.

* Refactor array and span libfuncs.

* Fix stuff.

* Fix stuff.

* Add debug programs.

* Fix bug pop front test bug.

* Fix slice libfunc errors.

* Fix more bugs.

* Fix keccak gas.

* Code cleanup.

* Add test for `array_snapshot_pop_back`.

* Add test for empty array cloning.

* Add realloc null comparison.

* Fix stuff.

* Fix slice start bound bug.

* Comment failing tests due to the felt to bool bug.

* Reenable ignored tests.

* Add tests we forgot to connect when combining all integration tests into a single binary.

* Fix details.

* Comment failing test.

+++ runtime/src/lib.rs
@@ -25,11 +25,11 @@ lazy_static! {
 pub unsafe extern "C" fn cairo_native__libfunc__debug__print(
     target_fd: i32,
     data: *const [u8; 32],
-    len: usize,
+    len: u32,
 ) -> i32 {
     let mut target = File::from_raw_fd(target_fd);
 
-    for i in 0..len {
+    for i in 0..len as usize {
         let data = *data.add(i);
 
         let value = Felt::from_bytes_le(&data);
